limiting damages to hydro
